# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: PACT Directory Service API
  version: 1.0.0
  description: >
    API for user authentication, organization management, inter-organization connections, 
    password reset, and proxying test execution to the conformance service.
servers:
  - url: http://localhost:3000/api
    description: Local
  - url: https://api.pact-directory.com
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]
    AuthTokenRequest:
      type: object
      properties:
        client_id: { type: string }
        client_secret: { type: string }
        network_key: { type: string }
      required: [client_id, client_secret, network_key]
    AuthTokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: Bearer }
      required: [access_token, token_type]
    SignupRequest:
      type: object
      properties:
        organizationName: { type: string }
        fullName: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 6 }
        confirmPassword: { type: string, format: password }
      required:
        - organizationName
        - fullName
        - email
        - password
        - confirmPassword
    SignupResponse:
      type: object
      properties:
        token: { type: string }
      required: [token]
    LoginRequest:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token: { type: string }
      required: [token]
    ForgotPasswordRequest:
      type: object
      properties:
        email: { type: string, format: email }
      required: [email]
    GenericMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
    ResetPasswordRequest:
      type: object
      properties:
        email: { type: string, format: email }
        token: { type: string }
        password: { type: string, minLength: 6 }
        confirmPassword: { type: string, minLength: 6 }
      required: [email, token, password, confirmPassword]
    VerifyResetTokenResponse:
      type: object
      properties:
        valid: { type: boolean }
        message: { type: string }
      required: [valid]
    OrganizationBase:
      type: object
      properties:
        id: { type: integer }
        organizationName: { type: string }
        organizationIdentifier: { type: string }
        organizationDescription: { type: string }
        solutionApiUrl: { type: string }
      required: [id, organizationName, organizationIdentifier]
    UserBase:
      type: object
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        role: { type: string }
      required: [fullName, email, role]
    Member:
      allOf:
        - { $ref: "#/components/schemas/UserBase" }
        - type: object
          properties:
            id: { type: integer }
            organizationId: { type: integer }
            organizationName: { type: string }
            organizationIdentifier: { type: string }
          required: [id]
    MyProfileResponse:
      type: object
      properties:
        userId: { type: integer }
        email: { type: string, format: email }
        organizationId: { type: integer }
        role: { type: string }
        policies: 
          type: array
          items: { type: string }
        fullName: { type: string }
        organizationName: { type: string }
        organizationIdentifier: { type: string }
        organizationDescription: { type: string }
        solutionApiUrl: { type: string }
        clientId: { type: string }
        clientSecret: { type: string }
        networkKey: { type: string }
        connectionRequests:
          type: object
          properties:
            sent:
              type: array
              items:
                type: object
                properties:
                  createdAt: { type: string, format: date-time }
                  status: { type: string }
                  organizationName: { type: string }
                  organizationId: { type: integer }
                required: [createdAt, status, organizationName, organizationId]
            received:
              type: array
              items:
                type: object
                properties:
                  id: { type: integer }
                  createdAt: { type: string, format: date-time }
                  status: { type: string }
                  organizationName: { type: string }
                  organizationId: { type: integer }
                required: [id, createdAt, status, organizationName, organizationId]
          required: [sent, received]
        connectedOrganizations:
          type: array
          items:
            type: object
            properties:
              organizationId: { type: integer }
              organizationName: { type: string }
              requestedAt: { type: string, format: date-time }
              createdAt: { type: string, format: date-time }
            required: [organizationId, organizationName, requestedAt, createdAt]
      required:
        - userId
        - email
        - organizationId
        - role
        - policies
        - fullName
        - organizationName
        - organizationIdentifier
        - solutionApiUrl
        - connectionRequests
        - connectedOrganizations
    OrganizationProfileResponse:
      type: object
      properties:
        id: { type: integer }
        organizationName: { type: string }
        organizationIdentifier: { type: string }
        organizationDescription: { type: string }
        networkKey: { type: string }
        solutionApiUrl: { type: string }
        parentId: { type: integer }
      required: [id, organizationName, organizationIdentifier, solutionApiUrl]
    OrganizationsMemberResponse:
      type: object
      properties:
        members:
          type: array
          items: { $ref: "#/components/schemas/Member" }
    CreateConnectionRequestRequest:
      type: object
      properties:
        requestedOrganizationId: { type: integer, description: "Requested organization ID" }
      required: [requestedOrganizationId]
    CreateConnectionRequestResponse:
      type: object
      properties:
        id: { type: integer }
      required: [id]
    ConnectionRequestActionRequest:
      type: object
      properties:
        requestId: { type: integer }
      required: [requestId]
    ConnectionCreatedResponse:
      type: object
      properties:
        message: { type: string, example: "Connection created successfully" }
      required: [message]
    RunTestCasesRequest:
      type: object
      properties:
        apiUrl: { type: string }
        clientId: { type: string }
        clientSecret: { type: string }
        version: { type: string }
        authBaseUrl: { type: string }
        scope: { type: string }
        resource: { type: string }
        audience: { type: string }
      required: [apiUrl, clientId, clientSecret, version]
    ExternalProxyResponse:
      description: Opaque response proxied from external conformance service.
      type: object
      additionalProperties: true
    AddUserToOrganizationRequest:
      type: object
      properties:
        fullName: { type: string, description: "Full name of the user", example: "John Doe" }
        email: { type: string, format: email, description: "Email address of the user", example: "john.doe@example.com" }
        role: { type: string, description: "Role of the user in the organization", example: "user" }
        password: { type: string, format: password, description: "Password for the user account", minLength: 6 }
        confirmPassword: { type: string, format: password, description: "Password confirmation" }
      required: [fullName, email, role, password, confirmPassword]
    UserContext:
      type: object
      properties:
        userId: { type: integer, description: "Unique identifier for the user", example: 123 }
        email: { type: string, format: email, description: "User's email address", example: "user@example.com" }
        organizationId: { type: integer, description: "ID of the organization the user belongs to", example: 456 }
        role: { type: string, description: "User's role in the organization", example: "user" }
        policies: 
          type: array
          items: { type: string }
          description: "List of policies/permissions assigned to the user"
          example: ["view-users", "edit-users"]
        status:
          type: string
          enum: [unverified, enabled, disabled, deleted]
      required: [userId, email, organizationId, role, policies, status]
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - bearerAuth: []
paths:
  /im/auth/token:
    post:
      summary: Issue inter-organization access token
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthTokenRequest' }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /directory/users/signup:
    post:
      summary: Sign up an organization and primary user
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: Signup successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignupResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /directory/users/login:
    post:
      summary: Login with email/password
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /directory/users/forgot-password:
    post:
      summary: Request password reset
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: Always returns generic success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMessage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /directory/users/reset-password:
    post:
      summary: Reset password with token
      tags: [Users]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: Password reset
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenericMessage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /directory/users/verify-reset-token/{token}:
    get:
      summary: Verify a password reset token
      tags: [Users]
      security: []
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Token verification result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyResetTokenResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
      x-notes: Controller currently reads req.params.token. Adjust controller or route.
  /directory/users/me:
    get:
      summary: Get current user profile with organization and connections
      tags: [Users]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: User profile with organization data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyProfileResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /directory/organizations/{id}:
    get:
      summary: Get organization profile by ID
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Organization profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationProfileResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
      x-notes: Controller uses req.params.id. Either change route to /directory/companies/profile/{id} or adapt controller to query.
  /directory/organizations:
    get:
      summary: List and search organizations
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: query
          required: false
          schema: { type: string }
          description: Search query to filter organizations by name
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OrganizationProfileResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /directory/organizations/{id}/users:
    get:
      summary: List all users in an organization
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: Organization ID
      responses:
        '200':
          description: List of users in the organization
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserBase' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Add a new user to an organization
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: Organization ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserToOrganizationRequest'
      responses:
        '201':
          description: User successfully added to organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /directory/organizations/{id}/connections:
    post:
      summary: List connections for an organization
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: Organization ID
      responses:
        '200':
          description: List of connections for the organization
          content:
            application/json:
              schema:
                type: array
                items: 
                  type: object
                  properties:
                    organizationId: { type: integer }
                    organizationName: { type: string }
                    requestedAt: { type: string, format: date-time }
                    createdAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /directory/organizations/create-connection-request:
    post:
      summary: Create a connection request between organizations
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateConnectionRequestRequest' }
      responses:
        '201':
          description: Connection request created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateConnectionRequestResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /directory/organizations/connection-request-action:
    post:
      summary: Accept connection request (creates connection between organizations)
      tags: [Organizations]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConnectionRequestActionRequest' }
      responses:
        '201':
          description: Connection created between organizations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConnectionCreatedResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /proxy/test:
    post:
      summary: Run test cases via proxy
      tags: [Proxy]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunTestCasesRequest' }
      responses:
        '200':
          description: Proxy response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalProxyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /proxy/test-results:
    get:
      summary: Get test run results
      tags: [Proxy]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: testRunId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Proxy response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalProxyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /proxy/test-runs:
    get:
      summary: List or search test runs
      tags: [Proxy]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: query
          required: false
          schema: { type: string }
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1 }
        - in: query
          name: pageSize
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Proxy response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalProxyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /directory/users/verify-email:
    post:
      tags:
        - Users
      summary: Verify email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: Email verification token
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /directory/users/resend-verification:
    post:
      tags:
        - Users
      summary: Resend email verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Verification email sent (if email exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
tags:
  - name: Authentication
  - name: Users
  - name: Organizations
  - name: Proxy